+++ src/services/Tx.ts
@@ -17,6 +17,7 @@ import { MangataGenericEvent } from "../types/MangataGenericEvent";
 import { MangataEventData } from "../types/MangataEventData";
 import { truncatedString } from "../utils/truncatedString";
 import { DepositXcmTuple, WithdrawXcmTuple } from "../types/AssetInfo";
+import { getWeightXTokens } from "../utils/getWeightXTokens";
 
 function serializeTx(api: ApiPromise, tx: SubmittableExtrinsic<"promise">) {
   if (!process.env.TX_VERBOSE) return "";
@@ -383,8 +384,13 @@ export class Tx {
         }
       };
 
+      const destWeightLimit = getWeightXTokens(
+        new BN(destWeight),
+        api.tx.xTokens.transferMultiasset
+      );
+
       await api.tx.xTokens
-        .transferMultiasset(asset, destination, new BN(destWeight))
+        .transferMultiasset(asset, destination, destWeightLimit)
         .signAndSend(account, {
           signer: txOptions?.signer,
           nonce: txOptions?.nonce
@@ -501,8 +507,13 @@ export class Tx {
       }
     };
 
+    const destWeightLimit = getWeightXTokens(
+      new BN("4000000000"),
+      api.tx.xTokens.transferMultiasset
+    );
+
     await turingApi.tx.xTokens
-      .transferMultiasset(asset, destination, new BN("4000000000"))
+      .transferMultiasset(asset, destination, destWeightLimit)
       .signAndSend(account, {
         signer: txOptions?.signer,
         nonce: txOptions?.nonce